+++ src/main/java/chess/domain/pieces/Bishop.java
@@ -2,6 +2,7 @@ package chess.domain.pieces;
 
 import chess.domain.Team;
 import chess.domain.math.Direction;
+import chess.domain.pieces.component.Name;
 
 import java.util.List;
