+++ src/test/java/chess/domain/pieces/KingTest.java
@@ -1,8 +1,8 @@
 package chess.domain.pieces;
 
-import chess.domain.File;
+import chess.domain.board.File;
 import chess.domain.Position;
-import chess.domain.Rank;
+import chess.domain.board.Rank;
 import chess.domain.Team;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
